requirements.txt
pytz~=2024.1
Scrapy~=2.11.1
Twisted~=24.3.0
APScheduler~=3.10.4
motor~=3.4.0
beanie~=1.25.0
pydantic~=2.7.0
uvicorn~=0.29.0
fastapi~=0.110.1
python-dotenv~=1.0.1
pymongo~=4.6.3
itemadapter~=0.8.0
pytube~=15.0.0
requests~=2.31.0
pillow~=10.3.0